// mobile/app/(main)/profileModal.tsx
import { Alert, Platform, ScrollView, StyleSheet, Text, TouchableOpacity, View } from 'react-native'
import React, { use, useEffect, useState } from 'react'
import { colors, spacingX, spacingY } from '@/constants/theme'
import { scale, verticalScale } from '@/utils/styling'
import { ScreenWrapper } from '@/components/ScreenWrapper'
import { Header } from '@/components/Header'
import { BackButtton } from '@/components/BackButton'
import { Avatar } from '@/components/Avatar'
import * as Icons from "phosphor-react-native"
import { Typo } from '@/components/Typo'
import { Input } from '@/components/Input'
import { useAuth } from '@/context/AuthContext'
import { UserDataProps } from '@/types'
import { Button } from '@/components/Button'
import { useRouter } from 'expo-router'
import { updateProfile } from '@/socket/socketEvents'
import * as ImagePicker from 'expo-image-picker';
import { uploadFileToCloudinary } from '@/services/imageService'

const Profile = () => {

    const router = useRouter();

    const [loading,setLoading] = useState(false);

    const { user, signOut, updateToken } = useAuth();

    const [userData,setUserData] = useState<UserDataProps>({
        name: "",
        email: "",
        avatar: null
    })

    const processUpdateProfile = (res: any) => {
        console.log("Got Result: ", res)
        setLoading(false);
        if(res.success) {
            updateToken(res.data.token);
            router.back();
        } else {
            Alert.alert("User", res.msg)
        }
    }

    useEffect(() => {
        updateProfile(processUpdateProfile);
        return () => {
            updateProfile(processUpdateProfile, true)
        }
    })

    useEffect(() => {
        setUserData({
            name: user?.name || "",
            email: user?.email || "",
            avatar: user?.avatar
        })
    }, [user])

    const handleUpdate = async () => {
        let { name, avatar } = userData;
        if(!name.trim()) {
            Alert.alert("User", "Please Enter Your Name");
            return;
        }
        let data = { name, avatar };
        if(avatar && avatar?.uri) {
            setLoading(true);
            const res = await uploadFileToCloudinary(avatar, "profiles")
            console.log("Result", res);
            if(res.success) {
                data.avatar = res.data
            } else {
                Alert.alert("User", res.msg);
                setLoading(false);
                return
            }
        }
        updateProfile(data);
    }

    const handleLogout = async () => {
        router.back();
        await signOut();
    }

    const onPickImage = async () => {
        let result = await ImagePicker.launchImageLibraryAsync({
            mediaTypes: ['images', 'videos'],
            aspect: [4, 3],
            quality: 1,
        });
        console.log(result);
        if (!result.canceled) {
            setUserData({ ...userData, avatar: result.assets[0] })
        }
    };

    const handleAlertSignOut = () => {
        Alert.alert("Confirm", "Are you sure want to logout ?", [
            {
                text: "Cancel",
                onPress: () => console.log("Cancel Logout"),
                style: "cancel"
            },
            {
                text: "Logout",
                onPress: () => handleLogout(),
                style: "destructive" 
            }
        ])
    }

    return (
        <ScreenWrapper
        isModal={true}
        >
            <View
            style={styles.container}
            >
                <Header
                title='Update Profile'
                leftIcon={
                    Platform.OS === "android" && <BackButtton color={colors.black} />
                }
                style={{ marginVertical: spacingY._15 }}
                />
                <ScrollView
                contentContainerStyle={styles.form}
                >
                    <View
                    style={styles.avatarContainer}
                    >   
                        <Avatar
                        uri={userData.avatar}
                        size={170}
                        />
                        <TouchableOpacity
                        style={styles.editIcon}
                        onPress={onPickImage}
                        >
                            <Icons.PencilIcon
                            size={verticalScale(20)}
                            color={colors.neutral800}
                            />
                        </TouchableOpacity>
                    </View>
                    <View
                    style={{gap: spacingY._20}}
                    >
                        <View
                        style={styles.inputContainer}
                        >
                            <Typo
                            style={{ paddingLeft: spacingX._10}}
                            >
                                Email
                            </Typo>
                            <Input
                            value={userData.email}
                            containerStyle={{
                                borderColor: colors.neutral350,
                                paddingLeft: spacingX._20,
                                backgroundColor: colors.neutral300
                            }}
                            onChangeText={(value) => setUserData({ ...userData, email: value })}
                            editable={false}
                            />
                        </View>
                        <View
                        style={styles.inputContainer}
                        >
                            <Typo
                            style={{ paddingLeft: spacingX._10}}
                            >
                                Name
                            </Typo>
                            <Input
                            value={userData.name}
                            containerStyle={{
                                borderColor: colors.neutral350,
                                paddingLeft: spacingX._20,
                            }}
                            onChangeText={(value) => setUserData({ ...userData, name: value })}
                            />
                        </View>
                    </View>
                </ScrollView>
            </View>
            <View
            style={styles.footer}
            >
                {!loading && (
                    <Button
                    style={{
                        backgroundColor: colors.rose,
                        height: verticalScale(56),
                        width: verticalScale(56)
                    }}
                    onPress={handleAlertSignOut}
                    >
                        <Icons.SignOutIcon
                        size={verticalScale(30)}
                        color={colors.white}
                        weight='bold'
                        />
                    </Button>
                )}
                <Button
                style={{ flex: 1 }}
                onPress={handleUpdate}
                loading={loading}
                >
                    <Typo
                    color={colors.black}
                    fontWeight={"700"}
                    >
                        Update
                    </Typo>
                </Button>
            </View>
        </ScreenWrapper>
    )
}

export default Profile

const styles = StyleSheet.create({
    inputContainer: {
        gap: spacingY._7
    },
    editIcon: {
        position: "absolute",
        bottom: spacingY._5,
        right: spacingY._7,
        borderRadius: 100,
        backgroundColor: colors.neutral100,
        shadowColor: colors.black,
        shadowOffset: { width: 0, height: 0 },
        shadowOpacity: 0.25,
        shadowRadius: 10,
        elevation: 4,
        padding: spacingY._7
    },
    avatar: {
        alignSelf: "center",
        backgroundColor: colors.neutral300,
        height: verticalScale(135),
        width: verticalScale(135),
        borderRadius: 200,
        borderWidth: 1,
        borderColor: colors.neutral500,
        //overflow: "hidden",
        //position: "relative"
    },
    avatarContainer: {
        position: "relative",
        alignSelf: "center"
    },
    form: {
        gap: spacingY._30,
        marginTop: spacingY._15
    },
    footer: {
        alignItems: "center",
        flexDirection: "row",
        justifyContent: "center",
        paddingHorizontal: spacingX._20,
        gap: scale(12),
        paddingTop: spacingY._15,
        borderTopColor: colors.neutral200,
        marginBottom: spacingY._10,
        borderTopWidth: 1
    },
    container: {
        flex: 1,
        justifyContent: "space-between",
        paddingHorizontal: spacingY._10,
        //paddingVertical: spacingY._30
    }
})